remove accessibility test file
