--- apps/main/e2e/accessibility.test.ts
@@ -1,82 +1,0 @@
-import { expect, test } from "./fixtures";
-
-test.describe("Keyboard-only Users Can", () => {
-  test("submit forms using Enter key", async ({ page }) => {
-    await page.goto("/learn");
-
-    const input = page.getByRole("textbox");
-    await input.fill("test prompt");
-    await page.keyboard.press("Enter");
-
-    // Should navigate to suggestions page
-    await expect(
-      page.getByRole("heading", { name: /course ideas for/i }),
-    ).toBeVisible({ timeout: 15_000 });
-  });
-
-  test("navigate course list items with Tab", async ({ page }) => {
-    await page.goto("/courses");
-
-    // Wait for courses to load
-    await expect(page.getByText("Machine Learning").first()).toBeVisible();
-
-    // Tab through course items
-    await page.keyboard.press("Tab");
-    await page.keyboard.press("Tab");
-    await page.keyboard.press("Tab");
-
-    // Should still be on courses page
-    await expect(
-      page.getByRole("heading", { name: /explore courses/i }),
-    ).toBeVisible();
-  });
-});
-
-test.describe("Screen Reader Support", () => {
-  test("command palette has dialog role", async ({ page }) => {
-    await page.goto("/");
-    await page.getByRole("button", { name: /search/i }).click();
-
-    const dialog = page.getByRole("dialog");
-    await expect(dialog).toBeVisible();
-  });
-
-  test("command palette has accessible title", async ({ page }) => {
-    await page.goto("/");
-    await page.getByRole("button", { name: /search/i }).click();
-
-    const dialog = page.getByRole("dialog");
-    // Dialog should have aria-label or aria-labelledby
-    const hasLabel = await dialog.evaluate(
-      (el) =>
-        el.hasAttribute("aria-label") || el.hasAttribute("aria-labelledby"),
-    );
-    expect(hasLabel).toBe(true);
-  });
-
-  test("search button indicates keyboard shortcut", async ({ page }) => {
-    await page.goto("/");
-
-    const searchButton = page.getByRole("button", { name: /search/i });
-    await expect(searchButton).toHaveAttribute("aria-keyshortcuts", /k/i);
-  });
-
-  test("form fields have accessible labels", async ({ page }) => {
-    await page.goto("/learn");
-
-    const input = page.getByRole("textbox");
-    // Input should have a label (via aria-label or associated label element)
-    const hasLabel = await input.evaluate((el) => {
-      const labelledby = el.getAttribute("aria-labelledby");
-      const label = el.getAttribute("aria-label");
-      const id = el.getAttribute("id");
-      const associatedLabel = id
-        ? document.querySelector(`label[for="${id}"]`)
-        : null;
-      return Boolean(
-        labelledby || label || associatedLabel || el.closest("label"),
-      );
-    });
-    expect(hasLabel).toBe(true);
-  });
-});
