+++ apps/main/e2e/learn.test.ts
@@ -58,3 +58,35 @@ test.describe("Course Suggestions", () => {
     ).toBeVisible();
   });
 });
+
+test.describe("Learn Form - Accessibility", () => {
+  test("submits form using Enter key", async ({ page }) => {
+    await page.goto("/learn");
+
+    const input = page.getByRole("textbox");
+    await input.fill("test prompt");
+    await page.keyboard.press("Enter");
+
+    await expect(
+      page.getByRole("heading", { name: /course ideas for/i }),
+    ).toBeVisible({ timeout: 15_000 });
+  });
+
+  test("form fields have accessible labels", async ({ page }) => {
+    await page.goto("/learn");
+
+    const input = page.getByRole("textbox");
+    const hasLabel = await input.evaluate((el) => {
+      const labelledby = el.getAttribute("aria-labelledby");
+      const label = el.getAttribute("aria-label");
+      const id = el.getAttribute("id");
+      const associatedLabel = id
+        ? document.querySelector(`label[for="${id}"]`)
+        : null;
+      return Boolean(
+        labelledby || label || associatedLabel || el.closest("label"),
+      );
+    });
+    expect(hasLabel).toBe(true);
+  });
+});
